move resource hash functions to internal
